# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: "Mechanics Shop API"
  description: "A comprehensive REST API for managing automotive service operations including customers, service tickets, mechanics, and inventory"
  version: "1.0.0"
  contact:
    name: "API Support"
    email: "support@mechanicshop.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://127.0.0.1:5000"
    description: "Development server"
  - url: "https://api.mechanicshop.com"
    description: "Production server"
components:
  securitySchemes:
    customerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Customer JWT Token"
    mechanicAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Mechanic JWT Token"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message describing what went wrong"
        details:
          type: object
          description: "Additional error details"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"

    # Customer Schemas
    CustomerLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "customer@example.com"
          description: "Customer's email address"
        password:
          type: string
          example: "password123"
          description: "Customer's password"
    CustomerRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "newcustomer@example.com"
          description: "Customer's email address (must be unique)"
        password:
          type: string
          example: "securepassword123"
          description: "Customer's password (min 6 characters)"
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "customer@example.com"
        password:
          type: string
          example: "hashedpassword"
    CustomerUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "updated@example.com"
        password:
          type: string
          example: "newpassword123"

    # Mechanic Schemas
...
    PaginatedCustomers:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 50
        pages:
          type: integer
          example: 5
        customers:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."

    # Health Check
...
paths:
...
  /customers/login:
    post:
      tags:
        - Authentication
      summary: "Customer Login"
      description: "Authenticate a customer and return a JWT token for accessing protected endpoints"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerLogin'
            examples:
              validLogin:
                summary: "Valid Customer Login"
                value:
                  email: "customer@example.com"
                  password: "password123"
      responses:
        200:
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  value:
                    token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        400:
          description: "Bad Request - Missing or invalid fields"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: "Unauthorized - Invalid credentials"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /customers/register:
    post:
      tags:
        - Customers
      summary: "Register New Customer"
      description: "Create a new customer account with email and password"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRegister'
            examples:
              newCustomer:
                summary: "New Customer Registration"
                value:
                  email: "newcustomer@example.com"
                  password: "securepassword123"
      responses:
        201:
          description: "Customer created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        400:
          description: "Bad Request - Missing required fields"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        409:
          description: "Conflict - Email already registered"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/all:
    get:
      tags:
        - Customers
      summary: "Get Paginated Customers"
      description: "Retrieve a paginated list of all customers. Rate limited to 10 requests per minute."
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: "Page number for pagination"
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: "Number of items per page"
      responses:
        200:
          description: "Customers retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomers'
        429:
          description: "Too Many Requests - Rate limit exceeded"
  /customers/{id}:
    get:
      tags:
        - Customers
      summary: "Get Customer by ID"
      description: "Retrieve a specific customer by their ID"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Customer ID"
      responses:
        200:
          description: "Customer found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        404:
          description: "Customer not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Customers
      summary: "Update Customer"
      description: "Update customer information (email and/or password)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Customer ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
            examples:
              updateCustomer:
                summary: "Update Customer Information"
                value:
                  email: "updated@example.com"
                  password: "newpassword123"
      responses:
        200:
          description: "Customer updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        404:
          description: "Customer not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Customers
      summary: "Delete Customer"
      description: "Permanently delete a customer account"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: "Customer ID"
      responses:
        200:
          description: "Customer deleted successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        404:
          description: "Customer not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Get Customer's Service Tickets"
      description: "Retrieve all service tickets for the authenticated customer. Requires customer authentication."
      security:
        - customerAuth: []
      responses:
        200:
          description: "Tickets retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    description:
                      type: string
                      example: "Brake repair needed"
        401:
          description: "Unauthorized - Missing or invalid token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== MECHANIC ENDPOINTS ==========
...